time-tracker/requirements.txt
# Core dependencies
PyYAML>=6.0
psutil>=5.9.0
pyobjc-framework-Quartz>=9.0; platform_system=='Darwin'
pyobjc-framework-AppKit>=9.0; platform_system=='Darwin'

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
mypy>=1.0.0
pylint>=2.15.0
